refactor: WishListJdbcRepository를 JPA 기반 WishListRepository로 전환

+++ src/main/java/gift/dto/WishSummaryResponseDto.java
@@ -4,7 +4,7 @@ import gift.domain.WishSummary;
 
 public record WishSummaryResponseDto(
         String productName,
-        int count
+        Long count
 ) {
     public static WishSummaryResponseDto from(WishSummary wishSummary) {
         return new WishSummaryResponseDto(
